fix(Run): Cambiado de double a short por problemas de memoria

+++ src/tech/Dilator.java
@@ -1,21 +1,21 @@
 package tech;
 
 public class Dilator {
-    private double[][] matrizOriginal;
-    private double[][] matrizDilatada;
+    private short[][] matrizOriginal;
+    private short[][] matrizDilatada;
     private int alto, ancho;
     private int figura;
 
-    public Dilator(double[][] matrizOriginal, int figura) {
+    public Dilator(short[][] matrizOriginal, int figura) {
         this.matrizOriginal = matrizOriginal;
         this.alto = matrizOriginal.length;
         this.ancho = matrizOriginal[0].length;
-        this.matrizDilatada = new double[alto][ancho];
+        this.matrizDilatada = new short[alto][ancho];
         this.figura = figura;
     }
     //Agrandar la matriz del canal en +1 en todas las direcciones
-    public double[][] expandMatrix(double[][] original) {
-        double[][] expandedMatrix = new double[alto + 2][ancho + 2];
+    public short[][] expandMatrix(short[][] original) {
+        short[][] expandedMatrix = new short[alto + 2][ancho + 2];
         // Llenar la nueva matriz con 0 (negro)
         for (int i = 0; i < alto + 2; i++) {
             for (int j = 0; j < ancho + 2; j++) {
@@ -32,15 +32,15 @@ public class Dilator {
     }
 
     // Método de dilatación secuencial usando la matriz expandida
-    public double[][] dilatacionSecuencial(int startRow, int endRow) {
+    public short[][] dilatacionSecuencial(int startRow, int endRow) {
         // Expandir la matriz original
-        double[][] matrizExpandida = expandMatrix(matrizOriginal);
+        short[][] matrizExpandida = expandMatrix(matrizOriginal);
         
         switch (figura) {
             case 1: // Figura 1: max[(x,y), (x-1,y), (x+1,y), (x,y-1), (x,y+1)]
                 for (int i = startRow; i <= endRow; i++)  {
                     for (int j = 1; j <= ancho; j++) {
-                        matrizDilatada[i - 1][j - 1] = Math.max(
+                        matrizDilatada[i - 1][j - 1] = (short)Math.max(
                             matrizExpandida[i][j],
                             Math.max(
                                 matrizExpandida[i - 1][j],
@@ -57,7 +57,7 @@ public class Dilator {
             case 2: // Figura 2: max[(x,y), (x-1,y), (x,y-1)]
                 for (int i = startRow; i <= endRow; i++)  {
                     for (int j = 1; j <= ancho; j++) {
-                        matrizDilatada[i - 1][j - 1] = Math.max(
+                        matrizDilatada[i - 1][j - 1] = (short)Math.max(
                             matrizExpandida[i][j],
                             Math.max(matrizExpandida[i - 1][j], matrizExpandida[i][j - 1])
                         );
@@ -68,7 +68,7 @@ public class Dilator {
             case 3: // Figura 3: max[(x,y), (x-1,y), (x,y+1)]
                 for (int i = startRow; i <= endRow; i++)  {
                     for (int j = 1; j <= ancho; j++) {
-                        matrizDilatada[i - 1][j - 1] = Math.max(
+                        matrizDilatada[i - 1][j - 1] =(short) Math.max(
                             matrizExpandida[i][j],
                             Math.max(matrizExpandida[i - 1][j], matrizExpandida[i][j + 1])
                         );
@@ -78,7 +78,7 @@ public class Dilator {
             case 4: // Figura 4: max[(x,y), (x-1,y), (x+1,y)]
                 for (int i = startRow; i <= endRow; i++)  {
                     for (int j = 1; j <= ancho; j++) {
-                        matrizDilatada[i - 1][j - 1] = Math.max(
+                        matrizDilatada[i - 1][j - 1] = (short)Math.max(
                             matrizExpandida[i][j],
                             Math.max(matrizExpandida[i - 1][j], matrizExpandida[i+1][j])
                         );
@@ -88,7 +88,7 @@ public class Dilator {
             case 5: // Figura 5: max[(x,y), (x,y-1)]
             for (int i = startRow; i <= endRow; i++)  {
                 for (int j = 1; j <= ancho; j++) {
-                    matrizDilatada[i - 1][j - 1] = Math.max(
+                    matrizDilatada[i - 1][j - 1] = (short)Math.max(
                         matrizExpandida[i][j],matrizExpandida[i][j-1]);
                     }
                 }
@@ -96,7 +96,7 @@ public class Dilator {
             case 6: // Figura 1: max[(x,y), (x-1,y), (x+1,y), (x,y-1), (x,y+1)]
                 for (int i = startRow; i <= endRow; i++)  {
                     for (int j = 1; j <= ancho; j++) {
-                        matrizDilatada[i - 1][j - 1] = Math.max(
+                        matrizDilatada[i - 1][j - 1] =(short) Math.max(
                             matrizExpandida[i][j],
                             Math.max(
                                 matrizExpandida[i - 1][j+1],
